Initial Upload
This is a simple Python desktop app that shows the current weather and local time for any city. Built with tkinter for the GUI, it uses Geopy and TimezoneFinder to detect the city's timezone, Pytz for accurate local time, OpenWeatherMap API to fetch real-time weather data.
It's a great mini-project for beginners learning how to work with APIs, timezones, and GUI applications in Python.